requirements.txt
feedparser
google-generativeai
requests
python-dotenv
line-bot-sdk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
